Fix up boxfactory completely.
Just need to install lua51, (and png).

Then a little cmake . && make

+++ gamelib/Algebra.h
@@ -4,7 +4,7 @@
 //
 // algebra.hpp/algebra.cpp
 //
-// Classes and functions for manipulating points, vectors, matrices, 
+// Classes and functions for manipulating points, vectors, matrices,
 // and colours.  You probably won't need to modify anything in these
 // two files.
 //
@@ -32,7 +32,7 @@ public:
     v_[1] = 0.0;
   }
   Point2D(double x, double y)
-  { 
+  {
     v_[0] = x;
     v_[1] = y;
   }
@@ -49,11 +49,11 @@ public:
     return *this;
   }
 
-  double& operator[](size_t idx) 
+  double& operator[](size_t idx)
   {
     return v_[ idx ];
   }
-  double operator[](size_t idx) const 
+  double operator[](size_t idx) const
   {
     return v_[ idx ];
   }
@@ -72,7 +72,7 @@ public:
     v_[2] = 0.0;
   }
   Point3D(double x, double y, double z)
-  { 
+  {
     v_[0] = x;
     v_[1] = y;
     v_[2] = z;
@@ -92,11 +92,11 @@ public:
     return *this;
   }
 
-  double& operator[](size_t idx) 
+  double& operator[](size_t idx)
   {
     return v_[ idx ];
   }
-  double operator[](size_t idx) const 
+  double operator[](size_t idx) const
   {
     return v_[ idx ];
   }
@@ -115,7 +115,7 @@ public:
     v_[2] = 0.0;
   }
   Vector3D(double x, double y, double z)
-  { 
+  {
     v_[0] = x;
     v_[1] = y;
     v_[2] = z;
@@ -135,11 +135,11 @@ public:
     return *this;
   }
 
-  double& operator[](size_t idx) 
+  double& operator[](size_t idx)
   {
     return v_[ idx ];
   }
-  double operator[](size_t idx) const 
+  double operator[](size_t idx) const
   {
     return v_[ idx ];
   }
@@ -207,7 +207,7 @@ inline Point3D operator -(const Point3D& a, const Vector3D& b)
   return Point3D(a[0]-b[0], a[1]-b[1], a[2]-b[2]);
 }
 
-inline Vector3D cross(const Vector3D& a, const Vector3D& b) 
+inline Vector3D cross(const Vector3D& a, const Vector3D& b)
 {
   return a.cross(b);
 }
@@ -240,7 +240,7 @@ public:
     v_[3] = 0.0;
   }
   Vector4D(double x, double y, double z, double w)
-  { 
+  {
     v_[0] = x;
     v_[1] = y;
     v_[2] = z;
@@ -263,11 +263,11 @@ public:
     return *this;
   }
 
-  double& operator[](size_t idx) 
+  double& operator[](size_t idx)
   {
     return v_[ idx ];
   }
-  double operator[](size_t idx) const 
+  double operator[](size_t idx) const
   {
     return v_[ idx ];
   }
@@ -292,28 +292,28 @@ public:
   {
     std::copy(other.v_, other.v_+16, v_);
   }
-  Matrix4x4(const Vector4D row1, const Vector4D row2, const Vector4D row3, 
+  Matrix4x4(const Vector4D row1, const Vector4D row2, const Vector4D row3,
              const Vector4D row4)
   {
-    v_[0] = row1[0]; 
-    v_[1] = row1[1]; 
-    v_[2] = row1[2]; 
-    v_[3] = row1[3]; 
+    v_[0] = row1[0];
+    v_[1] = row1[1];
+    v_[2] = row1[2];
+    v_[3] = row1[3];
 
-    v_[4] = row2[0]; 
-    v_[5] = row2[1]; 
-    v_[6] = row2[2]; 
-    v_[7] = row2[3]; 
+    v_[4] = row2[0];
+    v_[5] = row2[1];
+    v_[6] = row2[2];
+    v_[7] = row2[3];
 
-    v_[8] = row3[0]; 
-    v_[9] = row3[1]; 
-    v_[10] = row3[2]; 
-    v_[11] = row3[3]; 
+    v_[8] = row3[0];
+    v_[9] = row3[1];
+    v_[10] = row3[2];
+    v_[11] = row3[3];
 
-    v_[12] = row4[0]; 
-    v_[13] = row4[1]; 
-    v_[14] = row4[2]; 
-    v_[15] = row4[3]; 
+    v_[12] = row4[0];
+    v_[13] = row4[1];
+    v_[14] = row4[2];
+    v_[15] = row4[3];
   }
   Matrix4x4(double *vals)
   {
@@ -330,7 +330,7 @@ public:
   {
     return Vector4D(v_[4*row], v_[4*row+1], v_[4*row+2], v_[4*row+3]);
   }
-  double *getRow(size_t row) 
+  double *getRow(size_t row)
   {
     return (double*)v_ + 4*row;
   }
@@ -344,14 +344,14 @@ public:
   {
     return getRow(row);
   }
-  double *operator[](size_t row) 
+  double *operator[](size_t row)
   {
     return getRow(row);
   }
 
   const Matrix4x4 transpose() const
   {
-    return Matrix4x4(getColumn(0), getColumn(1), 
+    return Matrix4x4(getColumn(0), getColumn(1),
                       getColumn(2), getColumn(3));
   }
   Matrix4x4 invert() const;
@@ -364,7 +364,7 @@ public:
   {
     return begin() + 16;
   }
-		
+
 private:
   double v_[16];
 };
@@ -380,8 +380,8 @@ inline Matrix4x4 operator *(const Matrix4x4& aMatrix, const Matrix4x4& bMatrix)
     // Vector4D row = a.getRow(i);
     int row = i * 4;
 
-		
-    for(size_t j = 0; j < 4; ++j) {      
+
+    for(size_t j = 0; j < 4; ++j) {
       //ret[i][j] = a[i][0] * b[0][j] + a[i][1] * b[1][j] + a[i][2] * b[2][j] + a[i][3] * b[3][j];
       data[row + j] = a[row] * b[j] + a[row + 1] * b[4 + j] + a[row + 2] * b[8 + j] + a[row + 3] * b[12 + j];
     }
@@ -416,25 +416,25 @@ inline Vector3D transNorm(const Matrix4x4& M, const Vector3D& n)
 
 inline std::ostream& operator <<(std::ostream& os, const Matrix4x4& M)
 {
-  return os << "[" << M[0][0] << " " << M[0][1] << " " 
+  return os << "[" << M[0][0] << " " << M[0][1] << " "
             << M[0][2] << " " << M[0][3] << "]" << std::endl
-            << "[" << M[1][0] << " " << M[1][1] << " " 
+            << "[" << M[1][0] << " " << M[1][1] << " "
             << M[1][2] << " " << M[1][3] << "]" << std::endl
-            << "[" << M[2][0] << " " << M[2][1] << " " 
+            << "[" << M[2][0] << " " << M[2][1] << " "
             << M[2][2] << " " << M[2][3] << "]" << std::endl
-            << "[" << M[3][0] << " " << M[3][1] << " " 
+            << "[" << M[3][0] << " " << M[3][1] << " "
             << M[3][2] << " " << M[3][3] << "]";
 }
 
 class Colour
 {
 public:
-  Colour(double r, double g, double b)
+  Colour(float r, float g, float b)
     : r_(r)
     , g_(g)
     , b_(b)
   {}
-  Colour(double c)
+  Colour(float c)
     : r_(c)
     , g_(c)
     , b_(c)
@@ -453,26 +453,26 @@ public:
     return *this;
   }
 
-  double R() const 
-  { 
+  float R() const
+  {
     return r_;
   }
-  double G() const 
-  { 
+  float G() const
+  {
     return g_;
   }
-  double B() const 
-  { 
+  float B() const
+  {
     return b_;
   }
 
 private:
-  double r_;
-  double g_;
-  double b_;
+  float r_;
+  float g_;
+  float b_;
 };
 
-inline Colour operator *(double s, const Colour& a)
+inline Colour operator *(float s, const Colour& a)
 {
   return Colour(s*a.R(), s*a.G(), s*a.B());
 }
